refactor(evals): delete barrel runner.ts, extract runSuite, remove thin CLIs and npm scripts
