+++ package.json
@@ -77,10 +77,7 @@
     "check": "biome check --write .",
     "check:ci": "biome check .",
     "eval": "bun run src/cli/eval.ts",
-    "eval:all": "bun run src/cli/eval.ts",
     "eval:smoke": "bun run src/cli/eval.ts --smoke",
-    "eval:diff": "bun run src/cli/eval-diff.ts",
-    "collect": "bun run eval --collect",
     "ci:eval": "bun run eval --suite orchestrator-direct-execution --outputs-file /tmp/eval-outputs.json",
     "dev": "bun run build && opencode",
     "prepublishOnly": "bun run build",
